updating pythia program to handle QCD pT-binned, add batchsubmit

+++ mesonPt/pythia/setup.sh
@@ -2,14 +2,16 @@
 
 ver=pythia8243
 
-wget http://home.thep.lu.se/~torbjorn/pythia8/${ver}.tgz
-tar xf ${ver}.tgz
-rm ${ver}.tgz
+if [ ! -d ${ver} ]; then
+    wget http://home.thep.lu.se/~torbjorn/pythia8/${ver}.tgz
+    tar xf ${ver}.tgz
+    rm ${ver}.tgz
+fi
 
 cd $ver
 ./configure --with-root=$ROOTSYS
-
 make -j12
+export PYTHIA8DATA=`pwd`/share/Pythia8/xmldoc
 
 cd ..
